+++ app/src/main/java/com/sampietro/NaiveAAC/activities/Graphics/Images.java
@@ -1,5 +1,6 @@
 package com.sampietro.NaiveAAC.activities.Graphics;
 
+import static com.sampietro.NaiveAAC.activities.Settings.Utils.AdvancedSettingsDataImportExportHelper.dataProcess;
 import static com.sampietro.NaiveAAC.activities.Settings.Utils.AdvancedSettingsDataImportExportHelper.savBak;
 
 import android.content.Context;
@@ -13,6 +14,7 @@ import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.io.InputStreamReader;
 import java.nio.charset.StandardCharsets;
+import java.util.Objects;
 
 import io.realm.Realm;
 import io.realm.RealmObject;
@@ -91,7 +93,7 @@ public static void exporttoCsv(Context context, Realm realm){
     String header = AdvancedSettingsDataImportExportHelper.grabHeader(realm, "Images");
 
     // We write the header to file
-    AdvancedSettingsDataImportExportHelper.savBak(context, header,FILE_NAME );
+    savBak(context, header,FILE_NAME );
 
     // Now we write all the data corresponding to the fields grabbed above:
     //
@@ -104,7 +106,7 @@ public static void exporttoCsv(Context context, Realm realm){
             //
             dataP = taskitems.toString();
             // We process the data obtained and add commas and formatting:
-            dataP = AdvancedSettingsDataImportExportHelper.dataProcess(dataP);
+            dataP = dataProcess(dataP);
             // Workaround to remove the last comma from final string
             int total = dataP.length() - 1;
             dataP =  dataP.substring(0,total);
@@ -112,10 +114,10 @@ public static void exporttoCsv(Context context, Realm realm){
             // We write the data to file
             if (irrh == count-1)
             {
-                AdvancedSettingsDataImportExportHelper.savBak(context, dataP, FILE_NAME, true);
+                savBak(context, dataP, FILE_NAME, true);
             }
             else {
-                AdvancedSettingsDataImportExportHelper.savBak(context, dataP, FILE_NAME, false);
+                savBak(context, dataP, FILE_NAME, false);
             }
             irrh++;
         }
@@ -128,14 +130,18 @@ public static void exporttoCsv(Context context, Realm realm){
      *
      * @param context context
      * @param realm realm obtained from the activity by Realm#getDefaultInstance
+     * @param mode string import mode (Append or Replace)
      */
-    public static void importFromCsvFromInternalStorage(Context context, Realm realm)
+    public static void importFromCsvFromInternalStorage(Context context, Realm realm, String mode)
     {
-        // clear the table
-        RealmResults<Images> daCancellare = realm.where(Images.class).findAll();
-        realm.beginTransaction();
-        daCancellare.deleteAllFromRealm();
-        realm.commitTransaction();
+        if (Objects.equals(mode, "Replace"))
+            {
+                // clear the table
+                RealmResults<Images> daCancellare = realm.where(Images.class).findAll();
+                realm.beginTransaction();
+                daCancellare.deleteAllFromRealm();
+                realm.commitTransaction();
+            }
         //
         String rootPath = context.getFilesDir().getAbsolutePath();
         //
